requirements.txt
python-telegram-bot==20.8
dateparser==1.1.8
APScheduler==3.10.4
pytz==2023.3